backend/app/auth.py
from datetime import datetime, timedelta
from jose import JWTError, jwt
import bcrypt

SECRET_KEY = "SUPER_SECRET_KEY_FOR_ASSIGNMENT"
ALGORITHM = "HS256"

def hash_pass(password: str) -> str:
    # bcrypt requires bytes, so we encode the string to utf-8
    salt = bcrypt.gensalt()
    hashed_bytes = bcrypt.hashpw(password.encode('utf-8'), salt)
    # Return as a string to store in the database
    return hashed_bytes.decode('utf-8')

def verify_pass(plain: str, hashed: str) -> bool:
    # Encode both strings to bytes for comparison
    return bcrypt.checkpw(plain.encode('utf-8'), hashed.encode('utf-8'))

def create_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(hours=2)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)